Move Sentry setup to its own pack
This should ensure it gets executed earlier than the other
Javascript. Whilst it still _could_ execute later as its
`async`, I'd rather keep it like that for now to avoid a
performance hit.
