+++ app/webpacker/packs/application.js
@@ -1,14 +1,3 @@
-import * as Sentry from "@sentry/browser";
-
-const sentryConfig = document.querySelector("[data-sentry-dsn]")?.dataset;
-
-if (sentryConfig) {
-  Sentry.init({
-    dsn: sentryConfig.sentryDsn,
-    environment: sentryConfig.sentryEnvironment,
-  });
-}
-
 import '@stimulus/polyfills';
 import "core-js/stable";
 import "regenerator-runtime/runtime";
